The DICEModel Module

DICEModel — Module

DICEModel

Implementation of the DICE 2023 model

Notes:

Based on DICE2023-b-4-3-10.gms and included files (Nonco2-b-4-3-1.gms and FAIR-beta-4-3-1.gms)

Variable casing has been harmonized that all parameters and post-optimization computation have lower cases, and all optimization variables have upper case.

Module Index

DICEModel.RawParameters

DICEModel.run_dice

DICEModel.run_dice_scenario

DICEModel.@fields_to_vars

Detailed API

DICEModel.RawParameters — Type

Parameters

Parameters of the model. This structure hosts the "default" parameters that can then be modified using keyword arguments in the run_dice(pars) function.

First are defined some "raw" parameters, and then are defined some "computed" parameters (mostly arrays of ntsteps length). Both can be overriden using keyword arguments in the run_dice(pars) function.

DICEModel.run_dice — Method

run_dice(;optimizer=optimizer_with_attributes(Ipopt.Optimizer,"print_level" => 5), bounds=Dict{String,Tuple{String,String}}(),kwargs...)

Run the DICE model (currently v 2023).

This function runs the DICE model and returns the results as a named tuple.

Parameters

optimizer: The optimizer to use and eventually its options. Defaults: [optimizer_with_attributes(Ipopt.Optimizer,"print_level" => 5)].

bounds: A dictionary of equality or inequality constraints. Each constraint should be specified with the variable name as the key, and a two-elements tuple as the value. The first element is either "<=", ">=", or "==", and the second element is the right-hand side of the constraint (single value or a vector of ntimesteps length). Default: (empty dictionary). See the source code for the name of the variables of the model.

kwargs: Keyword arguments to override the default parameter values. Again, see the model source code for the name of the parameters.

Outputs

A named tuple containing the following fields: solved, status, times, tidx, the post_process computed values and the optimization variables.

Examples:

res = run_dice()
ECO2_opt = res.ECO2
plot(res.times[1:11] .+ 2020,ECO2_opt[1:11],ylim=(0,80), title="CO₂ emissions",ylabel="GtCO₂/yr",label="C/B optimal", markershape=:circle, markercolor=:white)

res_crazy = run_dice(optimizer=optimizer_with_attributes(Ipopt.Optimizer,"print_level" => 0), bounds = Dict("MIU"=>("==",1.0), "TATM"=>("<=",15), "Y" =>(">=",[fill(floatmin(Float64),10);fill(0.1,71)]), "ECO2" =>("<=",10000)), a2base = 0.01)

Notes

The bounds adds constraint to the problem, but do not substitute to hard-written bounds in the model. In particular, for the upper limit of the emissions controls, use the parameter miuup instead

bounds are always intended for the whole time steps. If you need bound for a subset of time steps (e.g. the first time step), you need to still compose your whole time array of the bound, using floatmin(Float64) or floatmax(Float64) as appropriate.

DICEModel.run_dice_scenario — Method

run_dice_scenario(scenario::String)

Run one of the "official" 10 scenarios in Nordhous's DICE 2023 model:

cbopt: The C/B optimal scenario

t2c: The temperature constrained to max 2 °C scenario

altdam: The alternative damage scenario

parisext: The Paris extended scenario

base: The base (current policies) scenario

r5: The scenario with discount rate of 5%

r4: The scenario with discount rate of 4%

r3: The scenario with discount rate of 3%

r2: The scenario with discount rate of 2%

r1: The scenario with discount rate of 1%

To run "your own" scenarios, use the function run_dice, where you can set the input parameters and constraints as you like.

Output

A named tuple containing the following fields: solved, status, times, tidx, the post_process computed values and the optimization variables.

DICEModel.@fields_to_vars — Macro

@fields_to_vars(t,x)

Utility macro to convert a struct fields to local variables (for readibility, so that we can write parx instead of using everywhere p.parx).